--- src/main/java/com/example/AgenciaTurismo/controller/JpaHotelController.java
@@ -1,43 +1,0 @@
-package com.example.AgenciaTurismo.controller;
-
-import com.example.AgenciaTurismo.dto.HotelDTO;
-import com.example.AgenciaTurismo.dto.response.ResponseDTO;
-import com.example.AgenciaTurismo.model.Hotel;
-import com.example.AgenciaTurismo.service.IJpaHotelService;
-import jakarta.validation.constraints.NotNull;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/v1/hotels")
-public class JpaHotelController {
-
-    @Autowired
-    IJpaHotelService service;
-
-    @GetMapping
-    public ResponseEntity<List<HotelDTO>> traerTodosLosHoteles() {
-        return new ResponseEntity<>(service.listarHotels(), HttpStatus.OK);
-    }
-
-    @PostMapping("/new")
-    public ResponseEntity<ResponseDTO> crearHotel(@RequestBody HotelDTO hotelDTO) {
-        return new ResponseEntity<>(service.createHotel(hotelDTO), HttpStatus.CREATED);
-    }
-
-    //UPDATE
-    @PostMapping("/edit/{id}")
-    public ResponseEntity<ResponseDTO> updateHotel(@PathVariable @NotNull Long id, @RequestBody HotelDTO hotelDTO) {
-        return new ResponseEntity<>(service.updateHotel(id, hotelDTO), HttpStatus.OK);
-    }
-
-    //DELETE
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<ResponseDTO> deleteHotel(@PathVariable @NotNull Long id) {
-        return new ResponseEntity<>(service.deleteHotel(id), HttpStatus.OK);
-    }
-}
