+++ src/main/java/com/example/AgenciaTurismo/repository/IHotelRepository.java
@@ -1,19 +1,12 @@
 package com.example.AgenciaTurismo.repository;
 
 import com.example.AgenciaTurismo.model.Hotel;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-import java.util.List;
+@Repository
+public interface IHotelRepository extends JpaRepository<Hotel, Long> {
 
-//public interface IHotelRepository{
-//    public List<Hotel> findAll();
-//
-//    //CREATE
-//    Hotel save(Hotel hotel);
-//
-//    //UPDATE
-//    Hotel update(Hotel hotel);
-//    //DELETE
-//    Hotel deleteHotel(String hotel);
-//}
-
-
+//    @Query("DELETE FROM Hotel h WHERE h.id = :hotelCode")
+//    void deleteByCode(@Param("hotelCode") String hotelCode);
+}
